Retrofit sendChatAudioResponse for batched actions
