+++ src/resolvers/chatResponse.ts
@@ -1,4 +1,8 @@
-import { setChatResponse, setChatSingleSelectResponse } from '../api'
+import * as uuid from 'uuid/v4'
+import { getUser, setChatResponse, setChatSingleSelectResponse } from '../api'
+import { s3 } from '../api/s3'
+import { AWS_CLAIMS_S3_BUCKET } from '../config'
+import { factory } from '../utils/log'
 import { setChatAudioResponse, setChatFileResponse } from './../api/index'
 import {
   MutationToSendChatAudioResponseResolver,
@@ -6,6 +10,10 @@ import {
   MutationToSendChatSingleSelectResponseResolver,
   MutationToSendChatTextResponseResolver,
 } from './../typings/generated-graphql-types'
+
+const logger = factory.getLogger('resolvers/chatResponse')
+
+const UPLOAD_OPTIONS = { partSize: 10 * 1024 * 1024, queueSize: 1 }
 
 export const sendChatTextResponse: MutationToSendChatTextResponseResolver = async (
   _root,
@@ -36,9 +44,37 @@ export const sendChatFileResponse: MutationToSendChatFileResponseResolver = asyn
 
 export const sendChatAudioResponse: MutationToSendChatAudioResponseResolver = async (
   _root,
-  { input },
+  { input: { file, globalId } },
   { getToken, headers },
 ) => {
   const token = getToken()
-  return setChatAudioResponse(token, headers, input)
+  const user = await getUser(token, headers)
+
+  const { createReadStream, filename, mimetype } = await file
+  const uploadedFileUrl: string = await new Promise((resolve, reject) => {
+    const key = `${uuid()}-${filename}`
+    const params = {
+      Bucket: AWS_CLAIMS_S3_BUCKET,
+      Key: key,
+      Body: createReadStream(),
+      ContentType: mimetype,
+      Metadata: {
+        MemberId: user.memberId,
+      },
+    }
+
+    s3.upload(params, UPLOAD_OPTIONS, (err, data) => {
+      if (err) {
+        logger.error('Got error when attempting to upload', err)
+        reject(err)
+        return
+      }
+
+      resolve(data.Location)
+    })
+  })
+  return setChatAudioResponse(token, headers, {
+    globalId,
+    url: uploadedFileUrl,
+  })
 }
