+++ src/schema.graphqls
@@ -38,7 +38,6 @@ type Mutation {
   emailSign: Boolean
   markMessageAsRead(globalId: ID!): Message!
   log(input: LoggingInput!): Boolean
-  uploadClaim(globalId: ID!, claim: Upload!): Boolean
 }
 
 type Subscription {
@@ -254,7 +253,7 @@ input ChatResponseBodyAudioInput {
 
 input ChatResponseAudioInput {
   globalId: ID!
-  body: ChatResponseBodyAudioInput!
+  file: Upload!
 }
 
 input CampaignInput {
