+++ src/typings/generated-graphql-types.ts
@@ -442,7 +442,6 @@ export interface Mutation {
   emailSign?: boolean;
   markMessageAsRead: Message;
   log?: boolean;
-  uploadClaim?: boolean;
 }
 
 export interface CampaignInput {
@@ -510,11 +509,7 @@ export interface ChatResponseBodyFileInput {
 
 export interface ChatResponseAudioInput {
   globalId: string;
-  body: ChatResponseBodyAudioInput;
-}
-
-export interface ChatResponseBodyAudioInput {
-  url: string;
+  file: Upload;
 }
 
 export interface TriggerClaimChatInput {
@@ -569,6 +564,10 @@ export enum OfferStatus {
 
 export interface SignEvent {
   status?: SignStatus;
+}
+
+export interface ChatResponseBodyAudioInput {
+  url: string;
 }
 
 /*********************************
@@ -1494,7 +1493,6 @@ export interface MutationTypeResolver<TParent = undefined> {
   emailSign?: MutationToEmailSignResolver<TParent>;
   markMessageAsRead?: MutationToMarkMessageAsReadResolver<TParent>;
   log?: MutationToLogResolver<TParent>;
-  uploadClaim?: MutationToUploadClaimResolver<TParent>;
 }
 
 export interface MutationToLogoutResolver<TParent = undefined, TResult = boolean> {
@@ -1637,14 +1635,6 @@ export interface MutationToLogArgs {
 }
 export interface MutationToLogResolver<TParent = undefined, TResult = boolean | null> {
   (parent: TParent, args: MutationToLogArgs, context: Context, info: GraphQLResolveInfo): TResult | Promise<TResult>;
-}
-
-export interface MutationToUploadClaimArgs {
-  globalId: string;
-  claim: Upload;
-}
-export interface MutationToUploadClaimResolver<TParent = undefined, TResult = boolean | null> {
-  (parent: TParent, args: MutationToUploadClaimArgs, context: Context, info: GraphQLResolveInfo): TResult | Promise<TResult>;
 }
 
 export interface SessionInformationTypeResolver<TParent = SessionInformation> {
